break(restructure): whole project
since we already break things - let's do it right!
