+++ README.md
@@ -43,6 +43,7 @@ OPTIONS
   -h, --help                     show CLI help
   -o, --out=out                  output directory
   -p, --preserve                 preserve output folder
+  -l, --localized                add localized types
   -s, --spaceId=spaceId          space id
   -t, --token=token              management token
   -v, --version                  show CLI version
