+++ src/renderer/field/render-prop-richtext.ts
@@ -1,6 +1,6 @@
 import {Field} from 'contentful';
-import {renderTypeUnion} from '../renderer';
-import {RenderContext} from './render-types';
+import {renderTypeUnion} from '../generic';
+import {RenderContext} from '../type';
 
 export const renderRichText = (field: Field, context: RenderContext): string => {
     context.imports.add({
