+++ src/renderer/generic/render-type-array.ts
