+++ src/renderer/generic/render-type-generic.ts
