--- src/type-renderer/index.ts
@@ -1,8 +1,0 @@
-export * from './render-types';
-export {renderPropAny} from './cf-render-prop-any';
-export {renderPropArray} from './cf-render-prop-array';
-export {renderPropLink} from './cf-render-prop-link';
-export {renderRichText} from './cf-render-prop-richtext';
-export {createDefaultContext} from './create-default-context';
-export {defaultRenderers} from './cf-default-renderers';
-export {DefaultContentTypeRenderer} from './cf-default-content-type-renderer';
